requirements.txt
fiona==1.10b2
geopandas==0.14.3
matplotlib==3.8.4
rasterio==1.3.9
shapely==2.0.4
scikit-learn==1.5.0
pytest-asyncio==0.23.6
tensorboard==2.16.2
python-dotenv==1.0.1
aiofiles==23.2.1
aiohttp_retry==2.8.3
timm
lightning==2.3.3
albumentations==1.4.0
tensordict==0.6.0
hydra-core==1.3.2
omegaconf==2.3.0
fvcore==0.1.5.post20221221
submitit>=1.5.1
pycocotools>=2.0.8
dataclasses-json>=0.6.7
segmentation-models-pytorch==0.3.3

